Corrected computation of noise mean/variance in IO

+++ dev/testing.py
@@ -29,7 +29,7 @@ if __name__=='__main__':
     print 'Number of unique masks:', data.num_masks
     
     kk = KK(data, max_iterations=1000)
-    kk.register_callback(SaveCluEvery(fname, shank, every=10))
+#     kk.register_callback(SaveCluEvery(fname, shank, every=10))
     kk.register_callback(MonitoringServer())
     
     def printclu_before(kk):
@@ -40,12 +40,12 @@ if __name__=='__main__':
         print 'Changed:', changed
         print kk.old_clusters[changed]
         print kk.clusters[changed]
-    kk.register_callback(printclu_before, 'start_EC_steps')
-    kk.register_callback(printclu_after, 'end_EC_steps')
+#     kk.register_callback(printclu_before, 'start_EC_steps')
+#     kk.register_callback(printclu_after, 'end_EC_steps')
     
-    if os.path.exists(fname+'.clu.'+str(shank)+'.flipflop'):
+    if False and os.path.exists(fname+'.clu.'+str(shank)+'.flipflop'):
         shutil.copy(fname+'.clu.'+str(shank)+'.flipflop', fname+'.clu.'+str(shank))
-    if os.path.exists(fname+'.clu.'+str(shank)):
+    if False and os.path.exists(fname+'.clu.'+str(shank)):
     #if False:
         print 'Loading clusters from file'
         clusters = loadtxt(fname+'.clu.'+str(shank), skiprows=1, dtype=int)
@@ -58,25 +58,25 @@ if __name__=='__main__':
     clusters = kk.clusters
     kk.reindex_clusters()
     
-    # num_to_show = 200
-    # 
-    # for cluster in xrange(kk.num_clusters_alive):
-    #     if cluster % 4 == 0:
-    #         figure()
-    #     maskimg = []
-    #     spikes = kk.get_spikes_in_cluster(cluster)
-    #     if len(spikes)>num_to_show:
-    #         spikes = sample(spikes, num_to_show)
-    #     for spike in spikes:
-    #         row = zeros(kk.num_features)
-    #         unmasked = data.unmasked[data.unmasked_start[spike]:data.unmasked_end[spike]]
-    #         row[unmasked] = data.masks[data.values_start[spike]:data.values_end[spike]]
-    #         maskimg.append(row)
-    #     if len(maskimg)==0:
-    #         continue
-    #     maskimg = array(maskimg)
-    #     subplot(2, 2, cluster%4 + 1)
-    #     imshow(maskimg, origin='lower left', aspect='auto', interpolation='nearest')
-    #     gray()
-    #     title(cluster)
-    # show()
+    num_to_show = 200
+     
+    for cluster in xrange(kk.num_clusters_alive):
+        if cluster % 4 == 0:
+            figure()
+        maskimg = []
+        spikes = kk.get_spikes_in_cluster(cluster)
+        if len(spikes)>num_to_show:
+            spikes = sample(spikes, num_to_show)
+        for spike in spikes:
+            row = zeros(kk.num_features)
+            unmasked = data.unmasked[data.unmasked_start[spike]:data.unmasked_end[spike]]
+            row[unmasked] = data.masks[data.values_start[spike]:data.values_end[spike]]
+            maskimg.append(row)
+        if len(maskimg)==0:
+            continue
+        maskimg = array(maskimg)
+        subplot(2, 2, cluster%4 + 1)
+        imshow(maskimg, origin='lower left', aspect='auto', interpolation='nearest')
+        gray()
+        title(cluster)
+    show()
